+++ app/context-personajes/page.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { useState, useEffect, createContext, ReactNode } from 'react'
 
 //Type  de personaje
