+++ app/types-ts/rick-and-morty-characters-info.ts
