--- src/layout.tsx
@@ -1,21 +1,0 @@
-import type { Metadata } from 'next'
-
-// These styles apply to every route in the application
-import './styles/globals.css'
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <html lang='en'>
-      <body>{children}</body>
-    </html>
-  )
-}
