+++ tailwind.config.js
@@ -2,7 +2,7 @@
 module.exports = {
   content: [
     // Or if using `src` directory:
-    './src/**/*.{js,ts,jsx,tsx,mdx}',
+    './app/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     extend: {},
